+++ example/lib/main.dart
@@ -1,4 +1,3 @@
-
 import 'package:anyline_tire_tread_plugin/anyline_tire_tread_plugin.dart';
 import 'package:anyline_tire_tread_plugin_example/env_info.dart';
 import 'package:anyline_tire_tread_plugin_example/home_screen.dart';