+++ lib/anyline_tire_tread_plugin.dart
@@ -1,7 +1,13 @@
 export 'package:anyline_tire_tread_plugin/src/tire_tread_plugin.dart';
 export 'src/scan_event.dart';
-export 'src/models/scan_options.dart';
 export 'src/enum/measurement_system.dart';
 export 'src/enum/scan_speed.dart';
+export 'src/enum/heatmap_style.dart';
+export 'src/enum/tire_side.dart';
 export 'src/models/tread_depth_result.dart';
+export 'src/models/tire_position.dart';
+export 'src/models/additional_context.dart';
+export 'src/models/scan_config.dart';
+export 'src/models/tire_tread_config.dart';
+export 'src/models/ui_config.dart';
 export 'src/experimental_flags.dart';