+++ lib/src/anyline_tire_tread_plugin_platform_interface.dart
@@ -1,4 +1,4 @@
-import 'package:anyline_tire_tread_plugin/src/models/scan_options.dart';
+import 'package:anyline_tire_tread_plugin/src/models/tire_tread_config.dart';
 import 'package:anyline_tire_tread_plugin/src/models/tread_depth_result.dart';
 import 'package:anyline_tire_tread_plugin/src/scan_event.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
@@ -36,7 +36,7 @@ abstract class AnylineTireTreadPluginPlatform extends PlatformInterface {
     throw UnimplementedError('initialize() has not been implemented.');
   }
 
-  Future<bool?> scan({required ScanOptions options}) {
+  Future<bool?> scan({required TireTreadConfig config}) {
     throw UnimplementedError('scan() has not been implemented.');
   }